+++ src/primitive/mod.rs
@@ -3,9 +3,10 @@
 //! For the meat of the actual array algorithms, see [`crate::algorithm`].
 
 mod defs;
-pub use defs::*;
+mod split;
 use ecow::EcoVec;
 use regex::Regex;
+pub use {defs::*, split::*};
 
 use core::str;
 use std::{
@@ -22,7 +23,6 @@ use std::{
 };
 
 use enum_iterator::{all, Sequence};
-use once_cell::sync::Lazy;
 use rand::prelude::*;
 use serde::*;
 
@@ -417,38 +417,6 @@ impl fmt::Display for FormatPrimitive {
     }
 }
 
-static ALIASES: Lazy<HashMap<Primitive, &[&str]>> = Lazy::new(|| {
-    [
-        (Primitive::Identity, &["id"] as &[_]),
-        (Primitive::Gap, &["ga"]),
-        (Primitive::Pop, &["po"]),
-        (Primitive::Fix, &["fx"]),
-        (Primitive::Box, &["bx"]),
-        (Primitive::IndexOf, &["idx"]),
-        (Primitive::ProgressiveIndexOf, &["pidx"]),
-        (Primitive::Switch, &["sw"]),
-        (Primitive::Stencil, &["st", "win"]),
-        (Primitive::Floor, &["flr", "flor"]),
-        (Primitive::Range, &["ran"]),
-        (Primitive::Partition, &["par"]),
-        (Primitive::Dup, &["dup"]),
-        (Primitive::Deshape, &["flat"]),
-        (Primitive::Ne, &["ne", "neq"]),
-        (Primitive::Eq, &["eq"]),
-        (Primitive::Lt, &["lt"]),
-        (Primitive::Le, &["le", "leq"]),
-        (Primitive::Gt, &["gt"]),
-        (Primitive::Ge, &["ge", "geq"]),
-        (Primitive::Utf8, &["utf", "utf__8"]),
-        (Primitive::First, &["fst"]),
-        (Primitive::Last, &["lst"]),
-        (Primitive::Slf, &["slf"]),
-        (Primitive::Select, &["sel"]),
-        (Primitive::Voxels, &["project"]),
-    ]
-    .into()
-});
-
 macro_rules! fill {
     ($ops:expr, $side:expr, $env:expr, $with:ident, $without_but:ident) => {{
         let env = $env;
@@ -577,242 +545,6 @@ impl Primitive {
     /// Check if this primitive is deprecated
     pub fn is_deprecated(&self) -> bool {
         self.deprecation_suggestion().is_some()
-    }
-    /// Get the short aliases for this primitive
-    pub fn aliases(&self) -> &'static [&'static str] {
-        ALIASES.get(self).copied().unwrap_or_default()
-    }
-    /// Try to parse a primitive from a name prefix
-    pub fn from_format_name(name: &str) -> Option<Self> {
-        if name.chars().any(char::is_uppercase) {
-            return None;
-        }
-        if name.len() < 2 {
-            return None;
-        }
-        static REVERSE_ALIASES: Lazy<HashMap<&'static str, Primitive>> = Lazy::new(|| {
-            ALIASES
-                .iter()
-                .flat_map(|(prim, aliases)| aliases.iter().map(|&s| (s, *prim)))
-                .collect()
-        });
-        if let Some(prim) = REVERSE_ALIASES.get(name) {
-            return Some(*prim);
-        }
-        if let Some(prim) = Primitive::non_deprecated().find(|p| p.name() == name) {
-            return Some(prim);
-        }
-        if let Some(prim) = Primitive::all().find(|p| p.glyph().is_none() && p.name() == name) {
-            return Some(prim);
-        }
-        if let Some(prim) = SysOp::ALL.iter().find(|s| s.name() == name) {
-            return Some(Primitive::Sys(*prim));
-        }
-        if name.len() < 3 {
-            return None;
-        }
-        let mut matching = Primitive::non_deprecated()
-            .filter(|p| p.glyph().is_some() && p.name().starts_with(name));
-        let res = matching.next()?;
-        let exact_match = res.name() == name;
-        (exact_match || matching.next().is_none()).then_some(res)
-    }
-    /// The list of strings where each character maps to an entire primitive
-    pub fn multi_aliases() -> &'static [(&'static str, &'static [(Primitive, &'static str)])] {
-        use Primitive::*;
-        &[
-            ("kork", &[(Keep, "k"), (On, "o"), (Rows, "r"), (Keep, "k")]),
-            ("rkok", &[(Rows, "r"), (Keep, "k"), (On, "o"), (Keep, "k")]),
-            ("awm", &[(Assert, "a"), (With, "w"), (Match, "m")]),
-            ("dor", &[(Div, "d"), (On, "o"), (Range, "r")]),
-            (
-                "pbbn",
-                &[(Partition, "p"), (Box, "b"), (By, "b"), (Ne, "n")],
-            ),
-            (
-                "ppbn",
-                &[(Partition, "p"), (Parse, "p"), (By, "b"), (Ne, "n")],
-            ),
-            (
-                "pibn",
-                &[(Partition, "p"), (Identity, "i"), (By, "b"), (Ne, "n")],
-            ),
-            ("kbn", &[(Keep, "k"), (By, "b"), (Ne, "n")]),
-            ("ftd", &[(Fork, "f"), (Take, "t"), (Drop, "d")]),
-            ("fdt", &[(Fork, "f"), (Drop, "d"), (Take, "t")]),
-            ("dnod", &[(Drop, "d"), (Neg, "n"), (On, "o"), (Drop, "d")]),
-            (
-                "adnoad",
-                &[
-                    (Anti, "a"),
-                    (Drop, "d"),
-                    (Neg, "n"),
-                    (On, "o"),
-                    (Anti, "a"),
-                    (Drop, "d"),
-                ],
-            ),
-            ("perf", &[(Dip, "p"), (Pop, "e"), (Under, "r"), (Now, "f")]),
-            ("wrench", &[(Sub, "wr"), (By, "en"), (Not, "ch")]),
-        ]
-    }
-    /// Look up a multi-alias from [`Self::multi_aliases`]
-    pub fn get_multi_alias(name: &str) -> Option<&'static [(Primitive, &'static str)]> {
-        Self::multi_aliases()
-            .iter()
-            .find(|(alias, _)| *alias == name)
-            .map(|(_, aliases)| *aliases)
-    }
-    /// Try to parse multiple primitives from the concatenation of their name prefixes
-    pub fn from_format_name_multi(name: &str) -> Option<Vec<(Self, &str)>> {
-        let mut indices: Vec<usize> = name.char_indices().map(|(i, _)| i).collect();
-        if indices.len() < 2 {
-            return None;
-        }
-        indices.push(name.len());
-        // Forward parsing
-        let mut prims = Vec::new();
-        let mut start = 0;
-        'outer: loop {
-            if start == indices.len() {
-                return Some(prims);
-            }
-            let start_index = indices[start];
-            for len in (2..=indices.len() - start).rev() {
-                let end_index = indices.get(start + len).copied().unwrap_or(name.len());
-                if end_index - start_index < 2 {
-                    continue;
-                }
-                let sub_name = &name[start_index..end_index];
-                // Normal primitive matching
-                if let Some(p) = Primitive::from_format_name(sub_name) {
-                    prims.push((p, sub_name));
-                    start += len;
-                    continue 'outer;
-                }
-                // 1-letter planet notation
-                if sub_name
-                    .strip_prefix('f')
-                    .unwrap_or(sub_name)
-                    .strip_suffix(['i', 'p', 'f'])
-                    .unwrap_or(sub_name)
-                    .chars()
-                    .all(|c| "gd".contains(c))
-                    && sub_name != "fi"
-                {
-                    for (i, c) in sub_name.char_indices() {
-                        let prim = match c {
-                            'f' if i == 0 => Primitive::Fork,
-                            'f' => Primitive::Fix,
-                            'g' => Primitive::Gap,
-                            'd' => Primitive::Dip,
-                            'i' => Primitive::Identity,
-                            'p' => Primitive::Pop,
-                            _ => unreachable!(),
-                        };
-                        prims.push((prim, &sub_name[i..i + 1]))
-                    }
-                    start += len;
-                    continue 'outer;
-                }
-                // Dip fix
-                if sub_name
-                    .strip_suffix('f')
-                    .unwrap_or(sub_name)
-                    .chars()
-                    .all(|c| c == 'd')
-                {
-                    for (i, c) in sub_name.char_indices() {
-                        let prim = match c {
-                            'd' => Primitive::Dip,
-                            'f' => Primitive::Fix,
-                            _ => unreachable!(),
-                        };
-                        prims.push((prim, &sub_name[i..i + 1]))
-                    }
-                    start += len;
-                    continue 'outer;
-                }
-                // Aliases
-                if let Some(ps) = Self::get_multi_alias(sub_name) {
-                    prims.extend(ps);
-                    start += len;
-                    continue 'outer;
-                }
-            }
-            break;
-        }
-        // Backward parsing
-        prims.clear();
-        let mut end = indices.len() - 1;
-        'outer: loop {
-            if end == 0 {
-                prims.reverse();
-                return Some(prims);
-            }
-            let end_index = indices[end];
-            for len in (2..=end).rev() {
-                let start_index = indices.get(end - len).copied().unwrap_or(0);
-                let sub_name = &name[start_index..end_index];
-                // Normal primitive matching
-                if let Some(p) = Primitive::from_format_name(sub_name) {
-                    prims.push((p, sub_name));
-                    end -= len;
-                    continue 'outer;
-                }
-                // 1-letter planet notation
-                if sub_name
-                    .strip_prefix('f')
-                    .unwrap_or(sub_name)
-                    .strip_suffix(['i', 'p', 'f'])
-                    .unwrap_or(sub_name)
-                    .chars()
-                    .all(|c| "gd".contains(c))
-                    && sub_name != "fi"
-                {
-                    for (i, c) in sub_name.char_indices().rev() {
-                        let prim = match c {
-                            'f' if i == 0 => Primitive::Fork,
-                            'f' => Primitive::Fix,
-                            'g' => Primitive::Gap,
-                            'd' => Primitive::Dip,
-                            'i' => Primitive::Identity,
-                            'p' => Primitive::Pop,
-                            _ => unreachable!(),
-                        };
-                        prims.push((prim, &sub_name[i..i + 1]))
-                    }
-                    end -= len;
-                    continue 'outer;
-                }
-                // Dip fix
-                if sub_name
-                    .strip_suffix('f')
-                    .unwrap_or(sub_name)
-                    .chars()
-                    .all(|c| c == 'd')
-                {
-                    for (i, c) in sub_name.char_indices().rev() {
-                        let prim = match c {
-                            'd' => Primitive::Dip,
-                            'f' => Primitive::Fix,
-                            _ => unreachable!(),
-                        };
-                        prims.push((prim, &sub_name[i..i + 1]))
-                    }
-                    end -= len;
-                    continue 'outer;
-                }
-                // Aliases
-                if let Some(ps) = Self::get_multi_alias(sub_name) {
-                    prims.extend(ps.iter().rev());
-                    end -= len;
-                    continue 'outer;
-                }
-            }
-            break;
-        }
-        None
     }
     /// Execute the primitive
     pub fn run(&self, env: &mut Uiua) -> UiuaResult {
@@ -2742,13 +2474,10 @@ mod tests {
         for (name, test) in [
             (
                 "from_format_name",
-                Primitive::from_format_name as fn(&str) -> Option<Primitive>,
+                PrimComponent::from_format_name as fn(&str) -> Option<PrimComponent>,
             ),
             ("from_format_name_multi", |name| {
-                Primitive::from_format_name_multi(name)
-                    .unwrap()
-                    .first()
-                    .map(|(prim, _)| *prim)
+                split_name(name).unwrap().first().map(|(prim, _)| *prim)
             }),
         ] {
             for prim in Primitive::non_deprecated() {
@@ -2761,7 +2490,7 @@ mod tests {
                     Some(_) => 4,
                 };
                 let short: String = prim.name().chars().take(char_test).collect();
-                assert_eq!(test(&short), Some(prim));
+                assert_eq!(test(&short), Some(prim.into()));
             }
             for prim in Primitive::non_deprecated() {
                 if matches!(
@@ -2778,7 +2507,7 @@ mod tests {
                 let short: String = prim.name().chars().take(char_test).collect();
                 assert_eq!(
                     test(&short),
-                    Some(prim),
+                    Some(prim.into()),
                     "{} does not format from {:?} with {}",
                     prim.format(),
                     short,
@@ -2792,18 +2521,24 @@ mod tests {
     #[test]
     fn from_multiname() {
         assert!(matches!(
-            &*Primitive::from_format_name_multi("rev").expect("rev"),
-            [(Primitive::Reverse, _)]
+            &*split_name("rev").expect("rev"),
+            [(PrimComponent::Prim(Primitive::Reverse), _)]
         ));
         assert!(matches!(
-            &*Primitive::from_format_name_multi("revrev").expect("revrev"),
-            [(Primitive::Reverse, _), (Primitive::Reverse, _)]
+            &*split_name("revrev").expect("revrev"),
+            [
+                (PrimComponent::Prim(Primitive::Reverse), _),
+                (PrimComponent::Prim(Primitive::Reverse), _)
+            ]
         ));
         assert!(matches!(
-            &*Primitive::from_format_name_multi("tabkee").unwrap(),
-            [(Primitive::Table, _), (Primitive::Keep, _)]
+            &*split_name("tabkee").unwrap(),
+            [
+                (PrimComponent::Prim(Primitive::Table), _),
+                (PrimComponent::Prim(Primitive::Keep), _)
+            ]
         ));
-        assert_eq!(Primitive::from_format_name_multi("foo"), None);
+        assert_eq!(split_name("foo"), None);
     }
 
     #[cfg(test)]